Start adding permissions

+++ backend/src/controllers/course/CourseAdministrationController.ts
@@ -10,14 +10,14 @@ import { sequelize } from "../../core/Sequelize";
 import { TrainingTypesBelongsToCourses } from "../../models/through/TrainingTypesBelongsToCourses";
 import { generateUUID } from "../../utility/UUID";
 import Validator, { ValidationTypeEnum } from "../../utility/Validator";
-
-// TODO: Move all course related things into this controller
+import PermissionHelper from "../../utility/helper/PermissionHelper";
+import { Transaction } from "sequelize";
+import { ForbiddenException } from "../../exceptions/ForbiddenException";
 
 /**
  * The ICourseBody Interface is the type which all requests that wish to create or update a course must satisfy
  */
 interface ICourseBody {
-    course_uuid: string;
     name_de: string;
     name_en: string;
     description_de: string;
@@ -29,27 +29,33 @@ interface ICourseBody {
 }
 
 /**
- * Validates and creates a new course based on the request
+ * Creates a new course within a given mentor group
  * @param request
  * @param response
  * @param next
  */
 async function createCourse(request: Request, response: Response, next: NextFunction) {
-    const transaction = await sequelize.transaction();
+    const transaction: Transaction = await sequelize.transaction();
+
     try {
         const user: User = response.locals.user;
         const body: ICourseBody = request.body as ICourseBody;
 
-        // Validator.validate(body, [
-        //     {
-        //         name: "course_uuid",
-        //         toValidate: [ValidationOptions.NON_NULL]
-        //     }
-        // ])
+        Validator.validate(body, {
+            name_de: [ValidationTypeEnum.NON_NULL],
+            name_en: [ValidationTypeEnum.NON_NULL],
+            description_de: [ValidationTypeEnum.NON_NULL],
+            description_en: [ValidationTypeEnum.NON_NULL],
+            active: [ValidationTypeEnum.NON_NULL],
+            self_enrol_enabled: [ValidationTypeEnum.NON_NULL],
+            training_type_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER],
+            mentor_group_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER],
+        });
 
-        if (!(await user.canManageCourseInMentorGroup(Number(body.mentor_group_id)))) {
-            response.sendStatus(HttpStatusCode.Forbidden);
-            return;
+        // Check if user is allowed to create a course in this mentor group
+        const mentorGroupID = Number(body.mentor_group_id);
+        if (!(await user.canManageCourseInMentorGroup(mentorGroupID))) {
+            throw new ForbiddenException("You don't have the required permission to create a course in this mentor group.", false);
         }
 
         const uuid = generateUUID();
@@ -72,7 +78,7 @@ async function createCourse(request: Request, response: Response, next: NextFunc
 
         await MentorGroupsBelongsToCourses.create(
             {
-                mentor_group_id: Number(body.mentor_group_id),
+                mentor_group_id: mentorGroupID,
                 course_id: course.id,
                 can_edit_course: true,
             },
@@ -106,13 +112,14 @@ async function createCourse(request: Request, response: Response, next: NextFunc
  * @param next
  */
 async function updateCourse(request: Request, response: Response, next: NextFunction) {
+    type ICourseBodyWithUUID = ICourseBody & { course_uuid: string };
+
     try {
         const user: User = response.locals.user;
-        const body: ICourseBody = request.body as ICourseBody;
+        const body: ICourseBodyWithUUID = request.body as ICourseBodyWithUUID;
 
         if (!(await user.canEditCourse(body.course_uuid))) {
-            response.sendStatus(HttpStatusCode.Forbidden);
-            return;
+            throw new ForbiddenException("You don't have the required permission to edit this course", false);
         }
 
         await Course.update(
@@ -140,12 +147,20 @@ async function updateCourse(request: Request, response: Response, next: NextFunc
 
 /**
  * Returns a list of all courses
- * @param request
+ * @param _request
  * @param response
+ * @param next
  */
-async function getAllCourses(request: Request, response: Response) {
-    const courses = await Course.findAll();
-    response.send(courses);
+async function getAllCourses(_request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        PermissionHelper.checkUserHasPermission(user, "lm.course.view");
+
+        const courses = await Course.findAll();
+        response.send(courses);
+    } catch (e) {
+        next(e);
+    }
 }
 
 /**
@@ -153,169 +168,239 @@ async function getAllCourses(request: Request, response: Response) {
  * This includes the initial training type used for this course
  * @param request
  * @param response
+ * @param next
  */
-async function getCourse(request: Request, response: Response) {
-    const params = request.params as { course_uuid: string };
+async function getCourse(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        PermissionHelper.checkUserHasPermission(user, "lm.course.view");
 
-    const course = await Course.findOne({
-        where: {
-            uuid: params.course_uuid,
-        },
-        include: [Course.associations.training_type],
-    });
+        const params = request.params as { course_uuid: string };
 
-    if (course == null) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        const course = await Course.findOne({
+            where: {
+                uuid: params.course_uuid,
+            },
+            include: [Course.associations.training_type],
+        });
+
+        if (course == null) {
+            response.sendStatus(HttpStatusCode.BadRequest);
+            return;
+        }
+
+        response.send(course);
+    } catch (e) {
+        next(e);
     }
-
-    response.send(course);
 }
 
 /**
  * Returns a list of users that are enrolled in this course.
  * @param request
  * @param response
+ * @param next
  */
-async function getCourseParticipants(request: Request, response: Response) {
-    const params = request.params as { course_uuid: string };
+async function getCourseParticipants(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        const params = request.params as { course_uuid: string };
+        if (!(await user.isMentorInCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not a mentor in this course.");
+        }
 
-    const course = await Course.findOne({
-        where: {
-            uuid: params.course_uuid,
-        },
-        include: [Course.associations.users],
-    });
+        const course = await Course.findOne({
+            where: {
+                uuid: params.course_uuid,
+            },
+            include: [Course.associations.users],
+        });
 
-    if (course == null) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        if (course == null) {
+            response.sendStatus(HttpStatusCode.BadRequest);
+            return;
+        }
+
+        response.send(course.users);
+    } catch (e) {
+        next(e);
     }
-
-    response.send(course.users);
 }
 
 /**
  * Removed a user from the specified course
  * @param request
  * @param response
+ * @param next
  */
-async function removeCourseParticipant(request: Request, response: Response) {
-    const params = request.params as { course_uuid: string };
-    const body = request.body as { user_id: number };
+async function removeCourseParticipant(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        const params = request.params as { course_uuid: string };
+        const body = request.body as { user_id: number };
 
-    const courseID = await Course.getIDFromUUID(params.course_uuid);
-    if (courseID == -1) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        if (!(await user.isMentorInCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not a mentor in this course.");
+        }
+
+        Validator.validate(body, {
+            user_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER],
+        });
+
+        const courseID = await Course.getIDFromUUID(params.course_uuid);
+        if (courseID == -1) {
+            response.sendStatus(HttpStatusCode.BadRequest);
+            return;
+        }
+
+        await UsersBelongsToCourses.destroy({
+            where: {
+                user_id: body.user_id,
+                course_id: courseID,
+            },
+        });
+
+        await TrainingRequest.destroy({
+            where: {
+                user_id: body.user_id,
+                course_id: courseID,
+            },
+        });
+
+        response.sendStatus(HttpStatusCode.NoContent);
+    } catch (e) {
+        next(e);
     }
-
-    await UsersBelongsToCourses.destroy({
-        where: {
-            user_id: body.user_id,
-            course_id: courseID,
-        },
-    });
-
-    await TrainingRequest.destroy({
-        where: {
-            user_id: body.user_id,
-            course_id: courseID,
-        },
-    });
-
-    response.sendStatus(HttpStatusCode.NoContent);
 }
 
 /**
  * Returns a list of mentor groups that are associated with this course
  * This does not include the permission such as 'can_edit'
+ *
+ * Restricted to mentors of this course.
  * @param request
  * @param response
+ * @param next
  */
-async function getCourseMentorGroups(request: Request, response: Response) {
-    const params = request.params as { course_uuid: string };
+async function getCourseMentorGroups(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        const params = request.params as { course_uuid: string };
 
-    const course = await Course.findOne({
-        where: {
-            uuid: params.course_uuid,
-        },
-        include: [
-            {
-                association: Course.associations.mentor_groups,
-                through: {
-                    attributes: ["can_edit_course", "createdAt"],
-                },
-                include: [
-                    {
-                        association: MentorGroup.associations.users,
-                        attributes: ["first_name", "last_name", "id"],
-                        through: {
-                            attributes: [],
-                        },
-                    },
-                ],
+        if (!(await user.isMentorInCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not a mentor in this course.");
+        }
+
+        const course = await Course.findOne({
+            where: {
+                uuid: params.course_uuid,
             },
-        ],
-    });
+            include: [
+                {
+                    association: Course.associations.mentor_groups,
+                    through: {
+                        attributes: ["can_edit_course", "createdAt"],
+                    },
+                    include: [
+                        {
+                            association: MentorGroup.associations.users,
+                            attributes: ["first_name", "last_name", "id"],
+                            through: {
+                                attributes: [],
+                            },
+                        },
+                    ],
+                },
+            ],
+        });
 
-    if (course == null) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        if (course == null) {
+            response.sendStatus(HttpStatusCode.BadRequest);
+            return;
+        }
+
+        response.send(course.mentor_groups);
+    } catch (e) {
+        next(e);
     }
-
-    response.send(course.mentor_groups);
 }
 
 /**
  * Adds a mentor group with the specified information to the course
  * @param request
  * @param response
+ * @param next
  */
-async function addMentorGroupToCourse(request: Request, response: Response) {
-    const body = request.body as { course_uuid: string; mentor_group_id: number; can_edit: boolean };
+async function addMentorGroupToCourse(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        const body = request.body as { course_uuid: string; mentor_group_id: number; can_edit: boolean };
 
-    const courseID = await Course.getIDFromUUID(body.course_uuid);
-    if (courseID == -1) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        Validator.validate(body, {
+            course_uuid: [ValidationTypeEnum.NON_NULL],
+            mentor_group_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER],
+            can_edit: [ValidationTypeEnum.NON_NULL],
+        });
+
+        if (!(await user.canEditCourse(body.course_uuid))) {
+            throw new ForbiddenException("You are not allowed to edit this course");
+        }
+
+        // If this returns -1, sequelize will complain due to a foreign constraint.
+        // Therefore, no need to check at this point.
+        const courseID = await Course.getIDFromUUID(body.course_uuid);
+
+        await MentorGroupsBelongsToCourses.create({
+            mentor_group_id: body.mentor_group_id,
+            course_id: courseID,
+            can_edit_course: body.can_edit,
+        });
+
+        response.sendStatus(HttpStatusCode.NoContent);
+    } catch (e) {
+        next(e);
     }
-
-    await MentorGroupsBelongsToCourses.create({
-        mentor_group_id: body.mentor_group_id,
-        course_id: courseID,
-        can_edit_course: body.can_edit,
-    });
-
-    response.sendStatus(HttpStatusCode.NoContent);
 }
 
 /**
  * Removes a mentor group from the specified course
  * @param request
  * @param response
+ * @param next
  */
-async function removeMentorGroupFromCourse(request: Request, response: Response) {
-    const params = request.params as { course_uuid: string };
-    const body = request.body as { mentor_group_id: number };
+async function removeMentorGroupFromCourse(request: Request, response: Response, next: NextFunction) {
+    try {
+        const user: User = response.locals.user;
+        const params = request.params as { course_uuid: string };
+        const body = request.body as { mentor_group_id: number };
 
-    Validator.validate(params, { course_uuid: [ValidationTypeEnum.NON_NULL] });
-    Validator.validate(body, { mentor_group_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER] });
+        Validator.validate(body, {
+            course_uuid: [ValidationTypeEnum.NON_NULL],
+            mentor_group_id: [ValidationTypeEnum.NON_NULL, ValidationTypeEnum.NUMBER],
+            can_edit: [ValidationTypeEnum.NON_NULL],
+        });
 
-    const courseID = await Course.getIDFromUUID(params.course_uuid);
-    if (courseID == -1) {
-        response.sendStatus(HttpStatusCode.BadRequest);
-        return;
+        if (!(await user.canEditCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not allowed to edit this course");
+        }
+
+        const courseID = await Course.getIDFromUUID(params.course_uuid);
+        if (courseID == -1) {
+            response.sendStatus(HttpStatusCode.BadRequest);
+            return;
+        }
+
+        await MentorGroupsBelongsToCourses.destroy({
+            where: {
+                course_id: courseID,
+                mentor_group_id: body.mentor_group_id,
+            },
+        });
+
+        response.sendStatus(HttpStatusCode.NoContent);
+    } catch (e) {
+        next(e);
     }
-
-    await MentorGroupsBelongsToCourses.destroy({
-        where: {
-            course_id: courseID,
-            mentor_group_id: body.mentor_group_id,
-        },
-    });
-
-    response.sendStatus(HttpStatusCode.NoContent);
 }
 
 /**
@@ -326,10 +411,14 @@ async function removeMentorGroupFromCourse(request: Request, response: Response)
  */
 async function getCourseTrainingTypes(request: Request, response: Response, next: NextFunction) {
     try {
+        const user: User = response.locals.user;
         const params = request.params as { course_uuid: string };
-        // TODO: Validate
 
-        const course = await Course.findOne({
+        if (!(await user.isMentorInCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not a mentor in this course.");
+        }
+
+        const course: Course | null = await Course.findOne({
             where: {
                 uuid: params.course_uuid,
             },
@@ -355,18 +444,26 @@ async function getCourseTrainingTypes(request: Request, response: Response, next
  */
 async function addCourseTrainingType(request: Request, response: Response, next: NextFunction) {
     try {
+        const user: User = response.locals.user;
         const params = request.params as { course_uuid: string };
         const body = request.body as { training_type_id: string };
-        // TODO: Validate
 
-        const course = await Course.findOne({ where: { uuid: params.course_uuid } });
+        Validator.validate(body, {
+            training_type_id: [ValidationTypeEnum.NON_NULL],
+        });
+
+        if (!(await user.canEditCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not allowed to edit this course.");
+        }
+
+        const course_id = await Course.getIDFromUUID(params.course_uuid);
 
         await TrainingTypesBelongsToCourses.create({
-            course_id: course?.id,
+            course_id: course_id,
             training_type_id: Number(body.training_type_id),
         });
 
-        response.sendStatus(HttpStatusCode.Ok);
+        response.sendStatus(HttpStatusCode.Created);
     } catch (e) {
         next(e);
     }
@@ -380,20 +477,28 @@ async function addCourseTrainingType(request: Request, response: Response, next:
  */
 async function removeCourseTrainingType(request: Request, response: Response, next: NextFunction) {
     try {
+        const user: User = response.locals.user;
         const params = request.params as { course_uuid: string };
         const body = request.body as { training_type_id: string };
-        // TODO: Validate
 
-        const course = await Course.findOne({ where: { uuid: params.course_uuid } });
+        Validator.validate(body, {
+            training_type_id: [ValidationTypeEnum.NON_NULL],
+        });
+
+        if (!(await user.canEditCourse(params.course_uuid))) {
+            throw new ForbiddenException("You are not allowed to edit this course.");
+        }
+
+        const course_id = await Course.getIDFromUUID(params.course_uuid);
 
         await TrainingTypesBelongsToCourses.destroy({
             where: {
-                course_id: course?.id,
+                course_id: course_id,
                 training_type_id: Number(body.training_type_id),
             },
         });
 
-        response.sendStatus(HttpStatusCode.Ok);
+        response.sendStatus(HttpStatusCode.NoContent);
     } catch (e) {
         next(e);
     }
